nanowire planner

+++ include/observer/http_observer.hpp
@@ -74,7 +74,7 @@ namespace acsr {
                 double step_size,
                 int min_steps, int max_steps, double sst_delta_near, double sst_delta_drain,
                 double optimization_distance,
-                int m, int n, double a, double b,
+                int m, int n, int a, double b,
                 const std::string& image_name
         ) override {
             start_time = system_clock::now();
@@ -307,6 +307,7 @@ namespace acsr {
                                                 const brynet::net::http::HttpSession::Ptr &session) {
                     (void) httpParser;
                     brynet::net::http::HttpResponse response;
+
                     std::string body = R"(
                 <!DOCTYPE html>
                 <html>
@@ -341,6 +342,7 @@ namespace acsr {
                                                                      true,
                                                                      false);
                     httpSession->send(frame);
+
                 };
 
                 brynet::net::wrapper::HttpListenerBuilder listenBuilder;
@@ -361,6 +363,7 @@ namespace acsr {
                                 brynet::net::http::HttpSessionHandlers &handlers) {
                             handlers.setHttpCallback(httpEnterCallback);
                             handlers.setWSCallback(wsEnterCallback);
+
                         })
                         .asyncRun();
