+++ include/observer/svg_observer.hpp
@@ -85,7 +85,7 @@ namespace acsr {
                 double step_size,
                 int min_steps,int max_steps, double sst_delta_near, double sst_delta_drain,
                 double optimization_distance,
-                int m,int n,double a,double b,
+                int m,int n,int a,double b,
                 const std::string& image_name
         ){
             original_image = createImage();
